requirements.txt
annotated-types==0.7.0
baml-py==0.90.2
pydantic==2.11.6
pydantic_core==2.33.2
python-dotenv==1.1.0
termcolor==3.1.0
typing-inspection==0.4.1
typing_extensions==4.14.0
chromadb==1.0.12
sentence-transformers==4.1.0